updateSnaphots: Fix edge case where meta data is missing but archive exists
Currently, the server has an up-to-date snapshot of REL1_29 and REL1_30,
but it seems those were not written to snapshotInfo.json.

Right now this means they will continue to be missing even when the script
has re-run many times since this happened because it will never re-create
the meta data if it finds the archive already exists.

Fix this by only skip archive+metadata creation if both archive
*and* meta data are up to date already. Previously it was skipping the
process based only on the archive existing, wrongly assuming that the
meta data also always exists in that case.